requirements.txt
colorama
pyOpenSSL